Fix executing current query/all queries in file
The elisp pulls out all defs in the file and prepends them when it
tries to dwim and run just the one query your cursor is currently on.
`C-c C-f` runs the whole file now, `C-c C-c` still does the
dwim-and-run-this-one-query. Improved error edge cases a little bit
too, if you happen to have the cursor positioned in a weird place (at
end of file, on a def) when you do `C-c C-c`.